+++ src/lib/pages/evm-contract-details/index.tsx
@@ -4,14 +4,26 @@ import { useRouter } from "next/router";
 import { useCallback, useEffect } from "react";
 
 import { AmpEvent, trackUseTab } from "lib/amplitude";
-import { useEvmConfig, useInternalNavigate } from "lib/app-provider";
+import {
+  useConvertHexAddress,
+  useEvmConfig,
+  useInternalNavigate,
+} from "lib/app-provider";
+import { AssetsSection } from "lib/components/asset";
 import { CustomTab } from "lib/components/CustomTab";
+import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import { CelatoneSeo } from "lib/components/Seo";
-import { InvalidState } from "lib/components/state";
+import { ErrorFetching, InvalidState } from "lib/components/state";
+import { useBalanceInfos } from "lib/services/bank";
+import {
+  useEvmCodesByAddress,
+  useEvmContractInfoSequencer,
+} from "lib/services/evm";
 import type { HexAddr20 } from "lib/types";
-import { isHexWalletAddress, truncate } from "lib/utils";
+import { is0x, isHexWalletAddress, truncate } from "lib/utils";
 
+import { EvmContractDetailsBytecode } from "./components/EvmContractDetailsBytecode";
 import { EvmContractDetailsOverview } from "./components/EvmContractDetailsOverview";
 import { EvmContractDetailsTop } from "./components/EvmContractDetailsTop";
 import { TabIndex, zEvmContractDetailsQueryParams } from "./types";
@@ -30,6 +42,20 @@ const EvmContractDetailsBody = ({
   tab,
 }: EvmContractDetailsBodyProps) => {
   const navigate = useInternalNavigate();
+  const { convertHexWalletAddress } = useConvertHexAddress();
+  const contractAddressBechAddr = convertHexWalletAddress(contractAddress);
+
+  const {
+    data: evmCodesByAddressData,
+    isLoading: isEvmCodesByAddressLoading,
+    isError: isEvmCodesByAddressError,
+  } = useEvmCodesByAddress(contractAddress);
+  const { data: evmContractInfoData, isLoading: isEvmContractInfoLoading } =
+    useEvmContractInfoSequencer(contractAddress);
+
+  const { totalData: totalAssets = 0 } = useBalanceInfos(
+    contractAddressBechAddr
+  );
 
   const handleTabChange = useCallback(
     (nextTab: TabIndex) => () => {
@@ -48,6 +74,14 @@ const EvmContractDetailsBody = ({
     },
     [contractAddress, tab, navigate]
   );
+
+  if (isEvmCodesByAddressLoading) return <Loading />;
+  if (
+    isEvmCodesByAddressError ||
+    !evmCodesByAddressData ||
+    is0x(evmCodesByAddressData.code)
+  )
+    return <ErrorFetching dataName="evm contract information" />;
 
   return (
     <>
@@ -70,7 +104,10 @@ const EvmContractDetailsBody = ({
             <CustomTab onClick={handleTabChange(TabIndex.Contract)}>
               Contract
             </CustomTab>
-            <CustomTab onClick={handleTabChange(TabIndex.Assets)}>
+            <CustomTab
+              onClick={handleTabChange(TabIndex.Assets)}
+              count={totalAssets}
+            >
               Assets
             </CustomTab>
             <CustomTab onClick={handleTabChange(TabIndex.Transactions)}>
@@ -78,12 +115,23 @@ const EvmContractDetailsBody = ({
             </CustomTab>
           </TabList>
           <TabPanels>
-            <TabPanel px={0}>
-              <EvmContractDetailsOverview />
+            <TabPanel p={0} pt={8}>
+              <EvmContractDetailsOverview
+                contractAddress={contractAddressBechAddr}
+                hash={evmContractInfoData?.hash}
+                sender={evmContractInfoData?.sender}
+                created={evmContractInfoData?.created}
+                isContractInfoLoading={isEvmContractInfoLoading}
+                onViewMoreAssets={handleTabChange(TabIndex.Assets)}
+              />
             </TabPanel>
-            <TabPanel px={0}>Contract</TabPanel>
-            <TabPanel px={0}>Assets</TabPanel>
-            <TabPanel px={0}>Transactions</TabPanel>
+            <TabPanel p={0} pt={8}>
+              <EvmContractDetailsBytecode code={evmCodesByAddressData.code} />
+            </TabPanel>
+            <TabPanel p={0}>
+              <AssetsSection address={contractAddressBechAddr} />
+            </TabPanel>
+            <TabPanel p={0}>Transactions</TabPanel>
           </TabPanels>
         </Tabs>
       </Stack>